Remove now unnecssary static_casts

+++ test/model/discount.cpp
@@ -6,8 +6,8 @@ TEST_CASE("Default discount is a no-op")
     model::Discount none;
     util::Amount zero(0), ten(10);
 
-    REQUIRE(static_cast<double>(none.applyTo(zero)) == 0.0);
-    REQUIRE(static_cast<double>(none.applyTo(ten)) == 10.0);
+    REQUIRE(none.applyTo(zero) == 0.0);
+    REQUIRE(none.applyTo(ten) == 10.0);
 }
 
 TEST_CASE("Discounts can be used on amounts")
